requirements.txt
numpy>=1.15.4
tensorflow>=1.12.0
pyyaml>=5.1.2=py36he774522_0